+++ zynq/examples/embassy/src/bin/dht22-open-drain-pins.rs
@@ -21,19 +21,9 @@ use zynq7000_hal::{
 };
 
 use zynq7000::Peripherals;
-use zynq7000_rt as _;
 
 // Define the clock frequency as a constant
 const PS_CLOCK_FREQUENCY: Hertz = Hertz::from_raw(33_333_300);
-
-/// Entry point (not called like a normal main function)
-#[unsafe(no_mangle)]
-pub extern "C" fn boot_core(cpu_id: u32) -> ! {
-    if cpu_id != 0 {
-        panic!("unexpected CPU ID {}", cpu_id);
-    }
-    main();
-}
 
 /// Try to talk to a DHT22 sensor connected at MIO0.
 const DHT22_AT_MIO0: bool = true;
@@ -41,8 +31,13 @@ const DHT22_AT_MIO0: bool = true;
 /// Open drain pin testing. MIO9 needs to be tied to MIO14.
 const OPEN_DRAIN_PINS_MIO9_TO_MIO14: bool = false;
 
+/// Entry point which calls the embassy main method.
+#[zynq7000_rt::entry]
+fn entry_point() -> ! {
+    main();
+}
+
 #[embassy_executor::main]
-#[unsafe(export_name = "main")]
 async fn main(_spawner: Spawner) -> ! {
     let mut dp = Peripherals::take().unwrap();
     l2_cache::init_with_defaults(&mut dp.l2c);