+++ zynq/examples/simple/src/main.rs
@@ -4,19 +4,9 @@
 
 use aarch32_cpu::asm::nop;
 use core::panic::PanicInfo;
-use zynq7000_rt as _;
 
-/// Entry point (not called like a normal main function)
-#[unsafe(no_mangle)]
-pub extern "C" fn boot_core(cpu_id: u32) -> ! {
-    if cpu_id != 0 {
-        panic!("unexpected CPU ID {}", cpu_id);
-    }
-    main();
-}
-
-#[unsafe(export_name = "main")]
-pub fn main() -> ! {
+#[zynq7000_rt::entry]
+fn main() -> ! {
     loop {
         nop();
     }